Extracted algorithms tests to separate test project
